backend/requirements.txt
apprise==0.9.2
APScheduler==3.7.0
asgiref==3.3.4
autopep8==1.5.7
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.2
django-apscheduler==0.5.2
django-filter==2.4.0
django-model-utils==4.1.1
djangorestframework==3.12.4
gunicorn==20.1.0
honcho==1.0.1
idna==2.10
Markdown==3.3.4
oauthlib==3.1.0
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pycparser==2.20
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
tzlocal==2.1
urllib3==1.26.4
